move dotproduct texture to new vectormath dir
