+++ include/slg/textures/vectormath/dotproduct.h
