+++ src/slg/CMakeLists.txt
@@ -421,7 +421,6 @@ set(SLG_CORE_SRCS
 	${LuxRays_SOURCE_DIR}/src/slg/textures/constfloat3.cpp
 	${LuxRays_SOURCE_DIR}/src/slg/textures/cloud.cpp
 	${LuxRays_SOURCE_DIR}/src/slg/textures/divide.cpp
-	${LuxRays_SOURCE_DIR}/src/slg/textures/dotproduct.cpp
 	${LuxRays_SOURCE_DIR}/src/slg/textures/dots.cpp
 	${LuxRays_SOURCE_DIR}/src/slg/textures/densitygrid.cpp
 	${LuxRays_SOURCE_DIR}/src/slg/textures/fbm.cpp
@@ -447,9 +446,10 @@ set(SLG_CORE_SRCS
 	${LuxRays_SOURCE_DIR}/src/slg/textures/subtract.cpp
 	${LuxRays_SOURCE_DIR}/src/slg/textures/texture.cpp
 	${LuxRays_SOURCE_DIR}/src/slg/textures/texturedefs.cpp
+	${LuxRays_SOURCE_DIR}/src/slg/textures/uv.cpp
+	${LuxRays_SOURCE_DIR}/src/slg/textures/vectormath/dotproduct.cpp
 	${LuxRays_SOURCE_DIR}/src/slg/textures/windy.cpp
 	${LuxRays_SOURCE_DIR}/src/slg/textures/wrinkled.cpp
-	${LuxRays_SOURCE_DIR}/src/slg/textures/uv.cpp
 	${LuxRays_SOURCE_DIR}/src/slg/utils/filenameresolver.cpp
 	${LuxRays_SOURCE_DIR}/src/slg/utils/pathdepthinfo.cpp
 	${LuxRays_SOURCE_DIR}/src/slg/utils/varianceclamping.cpp
