+++ src/slg/scene/parsetextures.cpp
@@ -43,7 +43,6 @@
 #include "slg/textures/constfloat3.h"
 #include "slg/textures/cloud.h"
 #include "slg/textures/divide.h"
-#include "slg/textures/dotproduct.h"
 #include "slg/textures/dots.h"
 #include "slg/textures/densitygrid.h"
 #include "slg/textures/fbm.h"
@@ -67,6 +66,7 @@
 #include "slg/textures/remap.h"
 #include "slg/textures/scale.h"
 #include "slg/textures/subtract.h"
+#include "slg/textures/vectormath/dotproduct.h"
 #include "slg/textures/windy.h"
 #include "slg/textures/wrinkled.h"
 #include "slg/textures/uv.h"
